TransferME/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
aiofiles
requests
spotipy
Pillow
